src/requirements/dev.txt
-r base.txt

# Testing
pytest>=8.3
pytest-django>=4.9
pytest-cov>=6.0
factory-boy>=3.3
faker>=33.0

# Code quality
ruff>=0.9
pre-commit>=4.1

# Debug
django-debug-toolbar>=5.0
ipython>=8.31
